Move UserDimContainer to an appropriate namespace
